+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotUserKeywordStatementId.java
@@ -8,6 +8,6 @@ import com.intellij.psi.PsiElement;
 public interface RobotUserKeywordStatementId extends RobotElement {
 
   @NotNull
-  List<RobotVariable> getVariableList();
+  List<RobotVariableDefinition> getVariableDefinitionList();
 
 }